pyproject.toml
[project]
name = "openai-sdr-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-agents>=0.13.2",
    "python-dotenv>=1.0.0",
]
